refactor: delete dublicate, improve code, increate rows
