+++ package.json
@@ -47,8 +47,6 @@
     "@emotion/cache": "^11.13.1",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
-    "@graphiql/react": "^0.26.2",
-    "@graphiql/toolkit": "^0.11.0",
     "@hookform/resolvers": "^3.9.0",
     "@mui/icons-material": "^5.16.7",
     "@mui/material": "^5.16.7",
@@ -58,8 +56,6 @@
     "firebase": "^10.13.1",
     "graphiql": "^3.7.1",
     "graphql": "^16.9.0",
-    "graphql-request": "^6.1.0",
-    "graphql-ws": "^5.16.0",
     "next": "^14.2.5",
     "next-international": "^1.2.4",
     "react": "^18.3.1",
