+++ src/components/custom-input/CustomInput.tsx
@@ -5,7 +5,7 @@ import type { ReactNode } from 'react';
 
 import styles from './CustomInput.module.scss';
 
-interface ICustomInput {
+interface CustomInputProps {
   label: string;
   variant: 'standard' | 'filled' | 'outlined';
   width?: string;
@@ -21,7 +21,7 @@ export default function CustomInput({
   value = '',
   onChange,
   onBlur,
-}: ICustomInput): ReactNode {
+}: CustomInputProps): ReactNode {
   return (
     <TextField
       label={label}
