+++ src/components/graphql-url-fieldset/GraphqlUrlFieldset.tsx
@@ -68,8 +68,8 @@ export default function GraphqlUrlFieldset(): ReactNode {
     toast.info('The request has been completed, look at the response body');
     window.history.pushState(null, '', window.location.href);
 
-    const requestsArray = (getStoredValue(LocalStorageKeys.URLS_RSS_REQUEST) as string[]) || [];
-    setStoredValue(LocalStorageKeys.URLS_RSS_REQUEST, [window.location.href, ...requestsArray]);
+    const requestsArray = (getStoredValue(LocalStorageKeys.REQUEST_LIST) as string[]) || [];
+    setStoredValue(LocalStorageKeys.REQUEST_LIST, [window.location.href, ...requestsArray]);
   };
 
   const handleSchemaRequest = async (): Promise<void> => {
